SLM_IoT_Control_Local/backend/requirements.txt
flask
ollama
python-dotenv
requests

langchain-chroma
langchain
langchain-community
langchain-ollama
pypdf
tiktoken

beautifulsoup4
lxml





